svg: Prevent code with fused instructions
Add a floating point conversion when multiplying when scaling
coordinates so that the multiplication does not get fused with an
adjacent addition or subtraction. This causes two instructions to be
generated, a MUL and a ADD/SUB, instead of a single FMADD or FMSUB.
A fused instruction has different rounding so ends up giving slightly
different results to two instructions. This was causing different values
in the SVGs generated between AMD64 and ARM64.

Link: https://go.dev/ref/spec#Floating_point_operators

+++ pkg/cli/svg/runtime.go
@@ -90,7 +90,14 @@ func (rt *GraphicsPlatform) transformY(y float64) float64 {
 }
 
 func (rt *GraphicsPlatform) scale(s float64) float64 {
-	return scaleFactor * s
+	// The float64 cast here acts as a barrier against a fused multiply/
+	// subtraction ARM64 instruction so that this multiply and a subtraction
+	// done by the caller of this function are not fused into a single ARM64
+	// instruction. This causes rounding differences when it is not used
+	// and causes the generated SVG files to have slight differences depending
+	// on which platform it is run.
+	// https://go.dev/ref/spec#Floating_point_operators
+	return float64(scaleFactor * s)
 }
 
 func (rt *GraphicsPlatform) nonDefaultAttr() Attr {
